fixed type of millis in notification class

+++ app/src/main/java/com/example/pillee/jetpackcompnavigation/model/Notification.kt
@@ -15,7 +15,7 @@ data class Notification (
 
 @Entity(tableName = "notification_table")
 data class NotificationDB(
-    @PrimaryKey val dateMillis: Int,
+    @PrimaryKey val dateMillis: Long,
     val pill: String,
     val userId: String
 )
